--- narrative_k8_router/__init__.py
@@ -1,10 +1,0 @@
-from fastapi import FastAPI
-
-
-def create_app() -> FastAPI:
-    # Create the FastAPI application instance
-    app = FastAPI()
-
-    # Configure the application
-    ###
-    return app
