+++ requirements.txt
@@ -3,6 +3,7 @@ fastapi==0.95.1
 python-dotenv==1.0.0
 httpx==0.24.0
 pytest==7.3.1
+pytest-cov==4.0.0
 kubernetes==26.1.0
 pydantic==1.10.7
 requests-mock==1.10.0
